Changing the folder structure of arc dfw
