+++ test/ttlens/unit_tests/test_lib.py
@@ -21,9 +21,10 @@ import os
 
 from ttlens.tt_lens_lib_utils import arc_write
 
-from ttlens.tt_arc_dbg_fw import ArcDebugLoggerFw, ArcDebugLoggerWithPmonFw
+from ttlens.arc_dfw.logger.tt_arc_logger_fw import ArcDebugLoggerFw
+from ttlens.arc_dfw.logger_with_pmon.tt_arc_logger_with_pmon_fw import ArcDebugLoggerWithPmonFw
 from ttlens.tt_lens_lib_utils import split_32bit_to_16bit
-from ttlens.tt_arc_dbg_fw_log_context import ArcDfwLogContextFromYaml, ArcDfwLogContextFromList
+from ttlens.arc_dfw.tt_arc_dbg_fw_log_context import ArcDfwLogContextFromList, ArcDfwLogContextFromYaml
 
 
 def invalid_argument_decorator(func):
