+++ ttlens/arc_dfw/logger_with_pmon/arc_dfw_logger_with_pmon.syms
@@ -1,0 +1,3 @@
+dfw_buffer_header: 00000548
+expandable_function: 00000550
+dfw_pmon_log: 00000304
